fix(): move createProviderClass from utils/general to utils/storyboard
