+++ packages/utils/src/general/index.ts
@@ -1,4 +1,3 @@
-export * from "./createProviderClass.js";
 export * from "./hasOwnProperty.js";
 export * from "./isObject.js";
 export * from "./JsonStorage.js";
