+++ packages/utils/src/storyboard/index.ts
@@ -1,3 +1,4 @@
+export * from "./createProviderClass.js";
 export * from "./expressions/index.js";
 export * from "./functions/index.js";
 export * from "./permissions.js";
